nuevos cambios(de usuario)

+++ backend/controllers/productosController/productoController.js
@@ -12,6 +12,7 @@ export const createProduct = async (req, res) => {
     }
     const producto = await Product.create({
       ...req.body,
+      userId: req.userId,
     });
     res.status(201).json({ msg: "Producto registrado" });
   } catch (error) {
@@ -24,20 +25,22 @@ export const createProduct = async (req, res) => {
 
 // Obtener todos los productos
 export const getProduct = async (req, res) => {
-    try {
-        const products = await Product.find().select("_id name price cost stock stockAmount barcode unit description");
-        res.status(200).json(products);
-    } catch (error) {
-        console.error("Error en getProduct:", error);
-        res.status(500).json({ message: "Error al obtener los productos", error });
-    }
+  try {
+    const products = await Product.find({ userId: req.userId }).select(
+      '_id name price cost stock stockAmount barcode unit description'
+    );
+    res.status(200).json(products);
+  } catch (error) {
+    console.error("Error en getProduct:", error);
+    res.status(500).json({ message: "Error al obtener los productos", error });
+  }
 };
 
 
 // Obtener un producto por ID
 export const getProductById = async (req, res) => {
     try {
-        const Product = await Product.findById(req.params.id);
+        const Product = await Product.findById({ _id: req.params.id, userId: req.userId });
         if (!Product) return res.status(404).json({ message: 'Producto no encontrado' });
         res.status(200).json(Product);
     } catch (error) {
@@ -55,7 +58,7 @@ export const getProductByBarcode = async (req, res) => {
       }
   
       // Buscamos el producto en la base de datos usando el barcode
-      const product = await Product.findOne({ barcode });
+      const product = await Product.findOne({ barcode, userId: req.userId });
   
       if (!product) {
         return res.status(404).json({ message: `Producto con barcode ${barcode} no encontrado` });
@@ -72,7 +75,9 @@ export const getProductByBarcode = async (req, res) => {
 // Actualizar un Producto
 export const updateProduct = async (req, res) => {
     try {
-        const updatedProduct = await Product.findByIdAndUpdate(req.params.id, req.body, { new: true });
+        const updatedProduct = await Product.findByIdAndUpdate( { _id: req.params.id, userId: req.userId },
+          req.body,
+          { new: true });
         if (!updatedProduct) return res.status(404).json({ message: 'Producto no encontrado' });
         res.status(200).json(Product);
     } catch (error) {
@@ -83,7 +88,7 @@ export const updateProduct = async (req, res) => {
 // Eliminar un producto
 export const deleteProduct = async (req, res) => {
     try {
-        const deletedProduct = await Product.findByIdAndDelete(req.params.id);
+        const deletedProduct = await Product.findByIdAndDelete({ _id: req.params.id, userId: req.userId });
         if (!deletedProduct) return res.status(404).json({ message: 'Producto no encontrado' });
         res.status(200).json({ message: 'Producto eliminado con éxito' });
     } catch (error) {
